Fix event test flakes.
Depending on the time of day the event tests are run, the generated
events may fall in the future. Orb's search endpoint *defaults* to a
`timeframe_end` of `now()`, so future events are excluded even if there
are ID matches. By explicitly setting the `timeframe_end` to one second
after the furthest future event we can ensure all events will always be
returned.