Update event amendment test to reflect new behavior.
Previously, when one amended an event, only the latest version of the
event was returned by the event search endpoint. That behavior has since
been revised by Orb, and now _all_ versions of events are returned.

Update the test to validate what we care about -- specifically, that the
amended event is returned by the search endpoint.